fix buffered produce batching
- calculating record numbers should return the current length of
buffered records rather than sub 1; the new record has not been added
yet
- fixes length accounting now that I know record header array length is
recorded as a varint

Also returns errors when a record cannot be enqueued rather than calling
the promise. This enables a Produce call to not hang if a user is not
ready for the promise to be called until after the Produce finishes.