commit WIP before another major overhaul
A bunch of work was done to make retrying easier. Well, now that I've
gotten this far, I realize that batches can just be buffered in what is
now known as the partition struct. This will also remove a bunch of
locking and the double map lookup that currently happens.

Instead, we can just assign partitions to brokers and throw everything
that is toppar now into the partition struct. On metadata refresh,
rather than overwriting the old partitions, we update the fields under a
lock.

It is worth a shot, at least.