generate: add multiple array levels
The BNF for CreatePartitions looked odd. Turns out the reason is because
the Assignments field ARRAY(INT32) actually meant that this was a multi
level array.

This adds parsing for multi level arrays, makes the code around arrays
clearer, and moves field type renaming to the struct itself rather than
after the array bracket (which is clearer imo).