kgo: minor optimizations; add max buffered messages
This commits a bunch of minor optimizations aimed at removing locks
(random partitioner case) or eliminating garbage.

The main optimization for existing code is to reuse batch slices. This
can be beneficial if spamming high volumes of messages.

The second optimization that is new to this committed code is around
retrying. Rather than wrapping the promise in another function, we call
a new defined function. A wrapper func would allocate for every message.