generate: add new type for FetchResponse record batch
It was quite the fun in the sun figuring this one out.

The BNF does not show that records is actually an array of record
batches. The linked doc describing RECORDS does not make this clear.
EVEN MORE CONFUSING is that RECORDS is used as NOT an array for
ProduceRequest

W
T
F

So, this adds hopefully the final type (since literally every other
request and response pair is more sane) to capture this logic.

The prior sized-struct I guess was ultimately unnecessary, but we can
keep it while leaving off the nullable bytes unnecessary field in
RecordBatch. Additionally, this tightens up the sized-struct to no
longer allow trailing nonsense on decode (not that we will be decoding
produce requests anyway).