kgo/errors: add ErrClientTooOld, unknown coordinator err
The unknown coordinator one will be cleaned up when that entire block of
errors is cleaned up.

ErrClientTooOld will likely only be returned when a person is using some
really cusom request type and satisfying the group coordinator interface
but we do not know how to parse the group ID out.