kgo: add auto group coordinator requesting
Group requests need to go to group coordinators. I originally thought to
just return the GroupID as a part of the group coordinator interface,
but describe and delete groups can work across multiple groups.

So, we need to split those requests. Easy enough, but the code is large.

We also want to retry group requests if the coordinator is not active,
loading, or moved. These all signify that our saved coordinator is out
of date.

This leaves a bunch of TODOs around retrying and the locking, but again,
the general logic is here.

I'll likely end up just using these functions for group requests when
writing the group consumer.