unify how metadata updates happen
When I began down the road of the group consumer, I quickly realized
that the metadata updating was jank.

There were TODOs in the producer to cleanup the metadata updating.

Ultimately, it makes sense to use the existing topicPartitions type for
consuming as well. The producer type does not need much state, nor does
the consumer.

This unifies metadata updating into a single file and cleans up a lot of
the flow of the code.

Update failures now also bump the try count on batches that are
currently in the backoff state.