remove another mutex in the produce path
The topicPartitions mutex existed so that we would not write to an old
broker after the records migrated to a new broker.

Since removeSource prevents that case, we can just rely on that and
instead switch to an atomic to remove a long held RWMutex.

This now gets the produce path down to basically one mutex.