move Fetch* from source to record, add iter, rename
The "primitives" of producing and consuming are now in the same file
for organization, and the file name is updated appropriately.

Adds an iterator for Fetches for what will likely be one of the most
common use cases of consuming: use all records for every poll, since
they are all from the same topic.

Likely requires adding an Errors function as well.