sticky: strengthen new scheme
The scheme could loop with a three way imbalance
A: 1 2     [interested in 1 2 3 4 5]
B: 3 4     [interested in 1 2 3 4 5]
C: 5       [interested in 1 2 3 4 5]
D: 6 7 8 9 [interested in 6 7 8 9]
E:         [interested in 6 7 8 9]
Cycle detection is needed.

This adds some light cycle detection but then also eliminates most
cycles (only allows there and back). This also strengthens the scheme so
that the above ordering will have no movements for A, B, C.