sticky: add some tests
I started adding all the sarama ones, but a lot of the complicated
scenarios I came up with more than cover most of the basic Java ones.

The multiple generations tests were worthwhile though, and exposed a
flaw in the current scheme. It's an easy fix. I don't take into account
stale partition consumers; I removed that thinking it was useless. Well,
if possible, we want to resticky first to those.