sticky: remove redundant partitionNums map
The partNums field was map[*topicPartition] from the switchover to using
slices and indices. On the way, I converted the map to a pointer to be
faster lookup.

This no longer needs to exist; the map is used rarely, so we can delete
the double work.

This gets back the throughput lost in the prior commit, now:
BenchmarkLarge-4             660           9076174 ns/op         2595759 B/op        501 allocs/op