generate: support non-nullable to nullable arrays by version
In the protocol, strings, bytes, and arrays can be null. Further,
some field can be non-null in one version and null in another.

Only arrays currently have transitioned from non-null to nullable,
and this has only happened for metadata requests and list offset
requests. This adds support for that

It would be pretty unfortunate to have strings or bytes transition
from non-null to nullable. This simple transitioning is a clear bleeding
of the Java type system into the protocol itself.