consuming: commit some major progress
Adds offset fetches and heartbets to group
- close to testing the group

Major direct consuming change:

Assigning is now similar to the group assign model where you pass
options, which means these two consuming modes are more likely to share
code in the future (see added option just below).

The options are easier to use than building a map.

This also adds immediate support for regex topic matching, which will be
shared with the group consumer soon and removes it from kcl.

Comparatively, the code is cleaner and easier to understand. There is a
little bit of cleanup, but nothing MAJOR left in here.

Also adds the first half of invalidating prior assignments on
reassigning partitions or groups.