finialize consuming reassigning, add leave group, cleanup consuming
This adds tracking which assignment sequence fetch requests are from
from when they begin using an offset thru when they are being unbuffered
for use by client users.

This allows us to easily invalidate prior fetches by bumping the seq and
checking it in a few spots.

With that in place, leaving a group is very easy.