consumer: fix memberID bug; config: move opt to producer
kcl had a problem where it kept continually rejoining after a long
period of time. Problem was that the member ID was not persisted past
one join session. Member IDs are now persisted until Join returns
UNKNOWN_MEMBER_ID.

Moves the request timeout opt to the producer due to the timeout field
only existing on the producer for requests used in this package.