remove a bunch of TODOs
Most of the TODOs are around config validation ensuring that numbers
make sense; instead, we'll just leave it up to API users to not do weird
incompatible numbers. It's not _our_ error if they do. We could add in
the validation in the future again, though.

Also allows buffering producing to partitions that have recoverable
errors, removing the wait loop in loading partitions for topic
producing.