reintroduce partition deletion handling; simplify source
If a topic is deleted and recreated with fewer partitions and then
deleted and recreated with more partitions, then
- the old partitions aren't cleaned up
- the new partitions aren't consumed since the direct consumer thinks it
is using them already and never fetches their offsets

This fixes that.

Cleans up and documents more clearly some of the consumer logic.

Also ignores internal topics from direct consuming regex.