complete KIP-480, add better produce partitioners
This makes private the partitioner interfaces right now until I figure
out a good interface strategy.

However, with the new API freedom, this removes the random partitioner
and adds in the two standard Java ones, which arguably are all that is
necessary, especially with KIP-480 support.

Moves the partitioner to the topicPartitions struct to avoid needing
global locks; adds a lock to the produce path but that seems
non-impacting in the 1.13 world even with defer (which is also why the
bufferRecord function switched to a defer in the prior commit).

Fixes a bug in using writable partitions as well while at it.

Very cool!