fix some nil pointer panics, rename variables
If a topic partition never loads successfully, we cannot use its the
sink or source. This means we must be careful with recordBuffer or
consumption to not access the sink or source if it is nil.

For variables, recBuf is nicer than recordBuffer, and c is nicer than
consumption (up for debate).