remove unneeded todo from README
The encoder already avoids writing the field if it does not exist. We
just wont provide an opt for the field. I added this because the KIP
changes some behavior if the field still exists on a new version, but I
misunderstood: this is if the field still exists because the client is
using an older version.