generate: explicitly die on unsupported tag types
KIP-482 is ambiguous as to whether strings / bytes will have their size
re-encoded in the opaque blob, so we will avoid those for now. I hope
"yes", because that would make it easier, but if the answer is no, I do
not want to accidentally have incorrect decoding in the future.

Structs and arrays of structs are more complicated; until a use case
exists for them, we will avoid it for now. For encoding, both of these
types likely require a two pass size calculation; that's not great.