draft of transactional consuming & producing
The public API is mostly there. I'm disappointed about the code
duplication in consume group, but with the types, it's likely
inevitable.

Work remains in fetching w.r.t. READ_COMMITTED, but I also still have
not figured out what the documentation means with fetching in that mode,
so that'll come as I figure out what is going on.

The AddPartitionsToTxn bit is a bit sketchy, but it may be the only real
way of doing this. Not quite sure how it'll play out yet. I guess,
testing time.