kgo: mostly finalize consuming committed
The aborted stuff isn't really explained clearly, but it's simpler in
implementation.

For producing, we need to actually set that a record is transactional
ourselves.

Moves TimestampType to an Attrs field in the record so that we can stuff
more bitfield info into a compact type.

Properly ignores aborted messages now in committed mode, as well as
control messages unless asked.