merge changes from bifurcation attempt for "sink"
I've decided that the bifurcation likely ultimately is abstracting for
the sake of "testing", when ultimately the the testing won't be good
anyway. Better testing can come in the form of integration tests.

This pulls back most changes in the "sink" package, although the produce
request changes are not pulled in yet.

By proxy, this also pulls in the "work" package and updates a bunch of
necessary changes in other files.

This doesn't build; the producer ID stuff does not yet exist, neither
does failing the producer ID. However, I have found that my current way
of doing the producer ID isn't entirely the best, so this is a good
opportunity to change it (there's no reason to get the producer ID
before buffering to a sink).