address KIP-360 and fix some seq num flaws
Some of this code, as the rest of the code in this package, needs to be
cleaned up a bit. However, there's not really great places to put some
of the new complicating piece.

First, the sequence number problem. I realized we actually cannot remove
any arbitrary batch that has not been sent because if we remove
something, future batches will have out of order sequence numbers. So,
this changes any place that removed a middle batch to remove / fail all
batches.

Second, this fixed a small issue with successfully finishing a batch 0
that actually could havev been failed hard by something unrelated. This
would be unlikely, but could happen.

For sequence numbers, this removes the number from every batch and
instead just has it on the recBuf. This makes accounting a little bit
harder, but in some ways makes it a little bit easier to reason about. I
wouldn't havev realized the sequence number problem with arbitrary
middle batches had I not been trying to solve resetting the number when
you have no knowledge of which batch did what.

This makes some errors in the transactional production a little bit
harder; it doesn't make sense to try to have any way to recover in some
cases, and doing so actually had some flaws (again, sequence number
problems).

In producer.go, this drastically simplifies initializing a producer ID,
and then adds some code to fail an ID per KIP-360. This doesn't try to
recover on failed IDs for idempotent producers. Instead, that is left
out in favor of the currently existing "stop on data loss" config knob.
The KIP itself doesn't really give any better guarantees if you're just
bumping the epoch and then resetting sequence numbers, so why not just
not do that. There's some pretty small risk though of resetting and then
having the reset be dropped as duplicate. I'm not sure that's worth
worrying about, though it would be easy to fix.