code cleanup & subtle issues in producing
For code cleanup, some things were moved around to simplify the code,
and some variable names were changed.
- makes the backoff logic simpler to understand, in the process fixing a
potential weird minor issue where a backoff could be triggered when it
shouldn't have been (due to racing between two atomics)
- fails all records when finishing batches on produce response if the
partition has a non-retriable error, because presumably this means kafka
did not write the records and thus we need to ensure our sequence
numbers are not broken
- makes handleRetryBatches logic simpler by ujst retrying on data loss
as well

Fixes two more even subtler issues in producer.go.
- for aborting records before ending a transaction, we need to ensure
that an AddPartitionsToTxn request would not go out following the call.
The code, as it was, allowed some (really unlikely) requests to
theoretically go out; we need to ensure they don't so that they do not
go out after we begin a new transaction.
- for hitting max buffered records, the original code was good, but then
adding ctx support introduced a flaw. This fixes that flaw by
un-counting our buffered record if we indeed do not buffer it.