add parsing string-ignore-nullable
KIP-559 changes a field that was always a string to be a
nullable-string, but there is no benefit to doing so other than nicer
code in the Kafka Java guts.

Since there is no benefit for the API, and there is no way to support a
string and a *string as a field in a struct, we will ignore the nullable
aspect of this field.