change JoinGroupResponse string-ignoreable-nullable to nullable-string-v#
Better to support the real protocol than ignore a signal. I'm fine with
some small breaking changes in the protocol if this happens more; the
protocol is a bit more special. We got lucky with slices being default
nullable, but other languages would not be so lucky. Well, we may as
well just do some minor breakages in the rare events they come up.