add BlockingCommitOffsets api for simpler committing
Writing the test proved that committing properly was a pain; we needed
to block in OnRevoke but didn't need to generally.

This adds a BlockingCommitOffsets API for times where blocking is
necessary, and then uses that API in the defaultRevoke. Now the test can
just use autocommitting, and consumers generally can just use
autocommitting.

This does not provide an analogous API for transactional committing
because transactions require much more care in general.