fix flexible request and response headers
My understanding was that the tags would trail the request body, because
in request bodies, tags trail the section they are in. That is, a
request header would wrap the body and the tags would come at the end.

Instead, apparently, headers work like HTML headers, i.e. they are
completely independent and the tags at the end of the header come before
the body.

Even more fun was to figure out why ApiVersions was so special.

Anyway, flexible versions works now.