consumer group: docs, races, and bugs
- some wording changes and clarifications in docs.
- memberID always needs locking to grab
  - further demonstrates it may be worth a better locking strategy in
the future such that things can only be accessed if a lock is grabbed
- allows updateUncommitted to rewind the head; this may happen if
ResetToCommitted rewound us, for example (or some other thing did)

And for the main bug:

fetchOffsets needs to set what we are committed to; we do not just want
to rely on CommitXyz. fetchOffsets fetches what *was* committed. If we
do not update our own committed map with this return, then
ResetToCommitted will erroneously rewind us.

After this, I'm not really getting test failures, so a cleanup pass is
worth it.