fix three more subtle logic errors
(1) When committing transactions, if the transaction offset commit
succeeds but the group is revoked before ending the transaction, we need
to reset to the _prior_ committed offsets. This adds a CommittedOffsets
function to enable this functionality, renames ResetToCommitted to
ResetOffsets, and renames Uncommitted to UncommittedOffsets for clarity.

This was not a problem until I made the lock tighter in session ending.
Before, the lock was held around both committing offsets and ending the
transaction. I made the lock tighter to further decrease risk of the two
requests taking longer than the rebalance timeout -- now it is just one
request. Odds were extremely slim for two requests to take so long
anyway, but may as well err on the side of safety.

(2) When assigning partitions, we need to duplicate the assignments map
that we are modifying.

When I previously "fixed" fetchOffsets by having it update the
uncommitted map, the updates to the uncommitted map used the same map
that was passed to assignPartitions. assignPartitions was deleting some
offsets (so that it could just remove what it did not need to list), but
this deleting breaks the fix in fetchOffsets.

This duplicates the assignments for simplicity; now, the fetchOffsets
fix from prior actually does what it intended to do.

(3) When ending a session, not erroring in commit txn offsets / end txn
does not mean we committed. This was an obvious oversight. Instead, we
need to return true only if both were nil _and_ we tried to commit.

---

Lastly, this updates some log lines.