consuming: fix subtle logic races
After this commit, my test looping for a few hours has not failed, so
that's a good thing.

The main gist of this comes down to group reassignments happening while
offsets were loading. If an offset was assigned that was loading, it'd
bump the seq, not trigger a new load, and the offset load would be
rejected.

The main fix it to track what is waiting to load, what is loading, and
to bump the seq's properly when necessary.

This would really only happen in cooperative consuming, which as it
turns out opens a whole can of worms for concurrency issues around
committing. Those should ideally be resolved now; using the client in a
_straightforward_ way should be correct. This doesn't attempt to solve
invalid usage.

Additionally, for rejoining groups, we now wait for the prior fetch loop
/ heartbeat loop to return before allowing rejoins.

Lastly, for leaving groups, this now waits for the manage loop to be
done. This is important because we do not want some weird race where the
member id we use is quickly invalidated. We want the last member ID ever
to use for leaving. This does make group leaving a bit weird, but eh.