generate: encode responses as well
The kmsg package may be beneficial to people writing "mock" brokers,
so this goes ahead and adds encoding responses.

Responses do not have the IsAdmin/IsGroup/IsTxn interfaces; those are
not necessary on the response side.

To support appending responses, this commit adds support for encoding
(basic) tags and appending nullable-string-v#+.