make tests better
This heavily refactors the tests so that group tests now also have the
same chaining that txn tests did, and we test all balancers.

The body size has been dropped by 100x because a large body was eating
all ram in the consumer side (due to it compressing well on the producer
side)

The chaining and more group leaves/joins show that the group test is
broken right now, so a fix for that is on the way.