fix test harness
Turns out that the residual group chain test was a test harness bug.
When leaving a group, we need to commit. However, when leaving early and
"not" committing, we also need to _not_ commit (since that's the
specific behavior we're trying to trigger).

Lastly, we need to remove offset order checking. There's a logic race,
especially so in group (not txn) chaining, where an etl could still be
working on its batch that it committed offsets for and then had revoked.
The work is fine, but we can't offset check at that point because
another consumer could have the partition and be working on a later
offset at the same time.

For transactions, the risk is much smaller but theoretically still
exists.

We have duplicate checking and all-consumed-produced checking, which
seems sufficient anyway. Order comes from Kafka, EOS comes from a
correct client. We're trying to ensure the client is correct.