config: switch to only allowing conn timeout overheads
Because we use internally coded requests, we cannot expose a function
that lets users set the timeout based on the request type. That will not
work with our internal requests.

For now, a simple workaround is to just provide the ability to change
the conn timeout overhead. I can't think of an obvious reason a person
would want different request timeouts than what the current function
does.