config: BrokerErrRetries => BrokerConnDeadRetries
The brokerErrRetries config knob was meant to detect specifically
ErrConnDead. I was thinking about how ErrNoDial should not cause us to
quit requests early, and then thinking about that more, we only want to
fail on repeated ErrConnDead. So, this switches the naming and logic
around to account for that.