add kerberos sasl
This needs to be tested. It's written from reading some sarama source
with some massive simplifications. Overall, there's about zero code
overlap.

The main surprising thing is that we write a final response but do not
read a challenge back (such as a final ok status). That's very weird.

Also, the ans1tools has a bug in it's length function and is way more
complicated than it needs to be; this rewrites that entirely.