bugfix, consumer group: properly return from heartbeat
For a consumer group, if it joined before anything was published to the
topic, then it would heartbeat and permanently set heartbeat to true.

The heartbeat bool never was reset, which meant that if we needed to
rejoin from rejoinCh, we wouldn't.

Resetting heartbeat ensures we rejoin as necessary if we discover new
topics or partitions and are the leader.

Secondly, let's avoid joining the group if we have no partitions for the
topic we want to consume. Otherwise, if we are a follower, our rejoin
will mean nothing because the metadata will not be updated.