consumer: fix two bugs
Turns out I can easily test what happens when partitions move around
with Kafka 2.4.0, because Kafka 2.4.0 introduced an alter partition
assignments command.

With testing, I found two bugs, both of which are relatively obvious in
hindsight.

First, a double lock; this is obvious when looking at it.

Second, when migrating a consumer, we need to reset the leader and the
preferred replica, even though that was not a part of these bug finds.
Otherwise, the preferred replica was the old leader in
processRespPartition, which means that we would perpetually set the
partition to unused at its old offset, trigger a metadata update, and
repeat the cycle.