producer: add more info lines (drop some debug to info)
The unknown topic wait messages are rare, so they may as well be info,
since it is pretty useful to know why we are not producing. Especially
so if the topic does not exist; before this, at level info, it would
just look like we are hanging forever.

Also adds some info messages to init producer id, which is also rare and
also good to know why it would be failing or that it is succeeding.