sink: ensure records are failed if the request attempt fails
Before, if a request hung / timed out / the connection died, then we
would not count that as a batch failure. We only counted batch failures
if the batch itself had a Kafka error in the produce response.

This switches to tracking all errors. This is generally what is expected
from the retry config toggle and ensures that we break out eventually if
we have limited retries in the face of repeated errors.