kmsg: break API for AppendRequest in partial support of KIP-590
KIP-590 adds two tags to the base request header. The client ID was
already added historically.

This breaks the API in favor of something that can be extended in a
non-breaking way in the future. All request headers have a base of the
bytes being appended to, the request being appended, and the correlation
ID of the request. Everything after is "new" from version 0.

KIP-590 adds two tags that should not be used by non-brokers, but kmsg
aims to support people implementing their own brokers.