v0.4.7: kmsg: breaking API in Read{RecordBatches,V{0,1}Messages}
As pointed out in #8, I had no crc validation on decode.
The most obvious place to add that validation (and length validation) is
in kmsg, which also makes this validation more broadly applicable to
those that do not want to use the kgo package.

This has been tested with the chaining tests, and v0 / v1 message sets
have been tested with kcl.