fetch request changes for KIP-595
This is **likely** not correct; I'll be testing this once 2.7.0 comes
out.

This is also slowly increasing the need for some default testing while
appending. With the way we append messages right now, all tags are
required to be set, as opposed to optionally not being sent on default
values.