generate: add support for "defaults"
This closes one of the last missing differences between my parser /
generator and the Java one. I'm not currently interested in closing the
missing "ignorable" aspect; I may be later, but unlikely.

Defaults for strings will be left out; they may be introduced later if a
an eventual tag comes around.