generate/DEFINITIONS: add defaults to mostly match Java
This change leaves out defaults on non-tagged fields where the Java
specified default is no different from the base default. Thus, this
leaves out defaults of 0 on numbers, defaults of null on nullable-*'s,
and defaults of "" on strings.

This also leaves out the default of true on MetadataRequests's
AllowAutoTopicCreation. I think there's even a KIP to deprecate this
default; regardless, this is not a super safe default to have.