kgo.Dialer: breaking change for easier API usage
By not including the network in the function passed to Dialer, the API
made it impossible to just pass a {net/tls}.Dialer.DialContext. Doing so
encouraged wrapper closures that would allocate a new dialer for every
dial.

The original API was chosen because I did not want to assume the network
(tcp or udp) that people wanted to use. This new change requires the
network to be tcp, but we can easily add a new option in the future,
DialerNetwork(string), that allows overriding the now-required tcp
network.

This new option makes it easier to use Dialer and matches other places
in the Go stdlib that take dialers (specifically, http.Transport's
DialContext)