generator: add support for enums
Enums have their own dedicated file, "enums",
and the syntax will be relatively simple:

// Comment
Name type (
  // Comment
  #: Word
)

and so on

enums can be used in every other file by prefixing a name with `enum-`,
so enum-Foo would refer to the Foo enum.

This will of course be a breaking change but should be a minor one and
will mostly affect kcl.

All enums have a default value of 0 for UNKNOWN.