broker: fix SaslAuthenticate, all: minor cleanliness
Turns out the hardcoded sasl authenticate key was wrong (36 vs 37). I'm
not sure how this worked on my local cluster with plain & scram. While
looking into that... this commit fixes that, avoids hardcoding keys in
broker (while keeping two keys hardcoded in client), and makes nicer
some log messages.

Lastly, for KIP-368, after looking at the PR, it looks like the first
step when reauthenticating is to handshake. So, this switches doSasl()
to sasl() to go to handshake first.