kgo: account for throttling
I avoided throttling previously because I considered it relatively
useless: we can send 5 produce requests concurrently, so how can we do
anything meaningful on the first throttle?

Well, we may as well at least try.

This uses reflect to account for throttling. The original reason to
avoid reflect was weak: binary bloat? Turns out it does not bloat much,
and checking for a field on response hardly adds any overhead
(determined through benchmarks), so this is fine.