client: handle retriable errors properly
For the retriable type, we never want to return the inner response
error. This is invalid; the response did not error. Instead, we want to
use that error to retry only when we can retry.

This properly separates out the response error and the inner retriable
error, and then for coordinator and admin requests, this properly only
returns the retriable error.

This bug has actually somewhat existed for a long time. Before, if we
were at retry limits, we would return our retriable error, which was
invalid.