generate: minor cleanup && support flexible non-top-level stuff
This makes non-top-level structs that have WithVersionField true more
unified with the top level encoders, and then adds an easier to reason
about regular expression to parse non-top-level structs. The regex now
parses flexible versions, which will be useful for the next commit.

This does shuffle around some field definitions in the encoding, but
that's small.