kgo: ensure we always have at least one shard
Non sharded requests always have one shard; that was a given before.
Sharded requests could avoid having any shards if the request actually
does not need to be issued. Rather than avoiding returning no shards,
this issues the request to any broker. This is not detrimental and makes
it easier to always know that there is at least one shard.