broker: add warning if connection dies on first read without sasl
If the successful reads is >0, or sasl is specified, then we log on
debug level. Otherwise, we log at warn. Hopefully this is not too noisy
if a broker is unreachable.

Kafka makes it impossible to determine if sasl is required if it is
missing; it just kills the connection. We can hint in a warning message
though to potentially look at missing sasl.

Closes #16.