consumer: fix allowUsable state; merge finishUsing
allowUsable originally had an atomic swap of the useState to 1, which
would allow a cursor to be consumed. This was accidentally deleted when
I removed the panics that I had to ensure state changes were correct.

As well, when I was ensuring state changes, I realized that the state
can be either used or unused in most cases when the state changes, so
ensuring the original state was a waste of time.

Lastly, the used state can basically be read as unusable, so having
three states as opposed to two was more complicated than necessary.

This removes finishUsing and fixes allowUsing.

Lastly, this moves on info message to debug.