generator: add support for UUID and fix non-top-level, no-encoding
UUID is simple to add support for; it's now just a primitive type.

For `not top level, no encoding` stuff, we need to add flexible
information sometimes. Particularly, for LeaderAndISR and
UpdateMetadata, the pulled-out-struct did not have the flexible
information bubbled down into it, which broke encoding / decoding. This
regex change and small logic change fixes that, which will be fixed in
the definitions in the next few commits.