describe quorum & fetch response minor changes
These two changes change nothing;
For describe quorum, this adds the flexible version to the extracted
common struct, but the common struct did not do anything different for
the flexible vs non-flexibole. This is a future proofing commit because
the flexible version should have been on the extracted struct.

For the fetch response, the generator added support for hex defaults;
this switches to the more obvious hex default.