kversion: switchup Versions API, finalize 2.7.0
As usual, a new release added new behavior: 2.7.0 has a gap in API keys
in the ApiVersions response.

This commit switches up the Versions type to a struct, which is a
breaking change but makes future changes unlikely to be breaking. The
gap is indicated with negative values.

This required small changes in the client guts, but I think the result
is easier to read.

Also removes an erroneous kversions bump for ApiVersions in tip.