producing: fail large messages more correctly
Before, a large message could fit within the max record batch
configuration, but could result in something too large for the wire.
This could result in a spin loop.

This should fix that, which also makes failing large messages more
accurate (before, we failed within 512 of the limit and only accounted
for key and value).